# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5f46db3167541c707d179879
  title: appvaa
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/states/{id}":
    delete:
      x-skaffolder-id: 5f46eb0067541c707d1799b9
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f46e55367541c707d179949
      x-skaffolder-resource: States
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - States
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f46eb0067541c707d1799b3
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f46e55367541c707d179949
      x-skaffolder-resource: States
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - States
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/States"
    post:
      x-skaffolder-id: 5f46eb0067541c707d1799b4
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f46e55367541c707d179949
      x-skaffolder-resource: States
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - States
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/States"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/States"
...
  "/students/{id}":
    delete:
      x-skaffolder-id: 5f46eac767541c707d179988
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f46e48867541c707d179924
      x-skaffolder-resource: Students
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Students
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f46eac767541c707d179982
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f46e48867541c707d179924
      x-skaffolder-resource: Students
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Students
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Students"
    post:
      x-skaffolder-id: 5f46eac767541c707d179981
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f46e48867541c707d179924
      x-skaffolder-resource: Students
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Students
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Students"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Students"
  "/students/findBy_class/{key}":
    get:
      x-skaffolder-id: 5f46eb1867541c707d1799cd
      x-skaffolder-name: findBy_class
      x-skaffolder-id-resource: 5f46e48867541c707d179924
      x-skaffolder-resource: Students
      x-skaffolder-crudAction: findBy_class
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_class
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_class/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_class
      tags:
        - Students
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Students"
  "/students/findBy_club/{key}":
    get:
      x-skaffolder-id: 5f46ead567541c707d179996
      x-skaffolder-name: findBy_club
      x-skaffolder-id-resource: 5f46e48867541c707d179924
      x-skaffolder-resource: Students
      x-skaffolder-crudAction: findBy_club
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_club
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_club/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_club
      tags:
        - Students
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Students"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    States:
      x-skaffolder-id: 5f46e55367541c707d179949
      x-skaffolder-id-db: 5f46db3167541c707d17987a
      x-skaffolder-id-entity: 5f46e55367541c707d179947
      x-skaffolder-url: /states
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5f46e55367541c707d179948
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - name
    Students:
      x-skaffolder-id: 5f46e48867541c707d179924
      x-skaffolder-id-db: 5f46db3167541c707d17987a
      x-skaffolder-id-entity: 5f46e48767541c707d179919
      x-skaffolder-url: /students
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        date_of_birth:
          type: integer
          x-skaffolder-id-attr: 5f46e48767541c707d17991a
          x-skaffolder-type: Date
          x-skaffolder-required: true
        facebook:
          type: string
          x-skaffolder-id-attr: 5f46e48767541c707d17991b
          x-skaffolder-type: String
        instagram:
          type: string
          x-skaffolder-id-attr: 5f46e48767541c707d17991c
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5f46e48767541c707d17991d
          x-skaffolder-type: String
          x-skaffolder-required: true
        password:
          type: string
          x-skaffolder-id-attr: 5f46e48767541c707d17991e
          x-skaffolder-type: String
          x-skaffolder-required: true
        position_secundary:
          type: number
          x-skaffolder-id-attr: 5f46e48767541c707d179920
          x-skaffolder-type: Number
          x-skaffolder-required: true
        position_usual:
          type: number
          x-skaffolder-id-attr: 5f46e48767541c707d17991f
          x-skaffolder-type: Number
          x-skaffolder-required: true
        steer:
          type: boolean
          x-skaffolder-id-attr: 5f46e48767541c707d179921
          x-skaffolder-type: Boolean
          x-skaffolder-required: true
        surname:
          type: string
          x-skaffolder-id-attr: 5f46e48767541c707d179922
          x-skaffolder-type: String
          x-skaffolder-required: true
        username:
          type: string
          x-skaffolder-id-attr: 5f46e48767541c707d179923
          x-skaffolder-type: String
          x-skaffolder-required: true
          x-skaffolder-unique: true
      x-skaffolder-relations:
        _class:
          x-skaffolder-id: 5f46e60967541c707d17995b
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5f46e48767541c707d179919
          x-skaffolder-ent2: 5f46e59767541c707d179951
        _club:
          x-skaffolder-id: 5f46e60967541c707d17995c
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5f46e48767541c707d179919
          x-skaffolder-ent2: 5f46e51a67541c707d17992b
      required:
        - _id
        - date_of_birth
        - name
        - password
        - position_secundary
        - position_usual
        - steer
        - surname
        - username
...